refactor(core): split 4 more oversized files to comply with 500 NLOC limit
- query_engine.rs (619→490): extract query_join.rs (136) for JOIN methods
- crud_bulk.rs (644→372): extract crud_bulk_raw.rs (293) for raw import path
- edge_concurrent/mod.rs (602→488): extract snapshot.rs (124) for CSR management
- query/mod.rs (651→284): extract options.rs (97) + wire pipeline_helpers.rs (302)

All files now under 500 NLOC. Zero public API changes. All tests pass.
